# public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sanduk API
  version: 1.0.0
  description: |
    REST API for Sanduk — a multi-user web app for organizing physical storage bins with QR codes.

    **Authentication**: Most endpoints require a JWT Bearer token obtained via `POST /auth/login`.
    Include it in the `Authorization: Bearer <token>` header.

    **Response envelopes**: All list endpoints return `{ results: T[], count: number }`.
    Error responses use `{ error: "CODE", message: "Human readable" }`.
servers:
  - url: /api
    description: API base path (proxied via nginx)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from POST /auth/login. Expires after 7 days.
  schemas:
...
    Bin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
        name:
          type: string
        area_id:
          type: string
          format: uuid
          nullable: true
        area_name:
          type: string
        items:
          type: array
          items:
            type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        icon:
          type: string
          description: PascalCase lucide icon name, empty string for default
        color:
          type: string
          description: Color preset key, empty string for no color
        short_code:
          type: string
          description: 6-character alphanumeric code for QR lookup
        created_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    CreateBinRequest:
      type: object
      required: [name, locationId]
      properties:
        name:
          type: string
          maxLength: 255
        locationId:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: string
          maxItems: 500
        notes:
          type: string
          maxLength: 10000
        tags:
          type: array
          items:
            type: string
          maxItems: 50
        areaId:
          type: string
          format: uuid
        icon:
          type: string
        color:
          type: string
    UpdateBinRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        areaId:
          type: string
          format: uuid
          nullable: true
        items:
          type: array
          items:
            type: string
          maxItems: 500
        notes:
          type: string
          maxLength: 10000
        tags:
          type: array
          items:
            type: string
          maxItems: 50
        icon:
          type: string
        color:
          type: string
    AddTagsRequest:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            type: string
          maxItems: 50

    # ── Photo ──────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Authentication required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Not a member of this location
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Resource not found
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Name is required
...
paths:
...
  /bins:
    post:
      tags: [Bins]
      summary: Create a bin
      description: |
        Creates a new bin with an auto-generated 6-character short code.
        The short code uses a charset excluding ambiguous characters (0/O, 1/l/I).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBinRequest'
      responses:
        '201':
          description: Bin created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Bins]
      summary: List bins for a location
      description: Returns non-deleted bins. Requires location_id query parameter.
      parameters:
        - name: location_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of bins
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bin'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /bins/trash:
    get:
      tags: [Bins]
      summary: List soft-deleted bins
      description: Returns bins with non-null deleted_at. Auto-purges after 30 days.
      parameters:
        - name: location_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of trashed bins
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bin'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /bins/lookup/{shortCode}:
    get:
      tags: [Bins]
      summary: Look up a bin by short code
      description: Used for manual QR code entry. Returns the bin matching the 6-character code.
      parameters:
        - name: shortCode
          in: path
          required: true
          schema:
            type: string
            minLength: 6
            maxLength: 6
      responses:
        '200':
          description: Bin found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /bins/{id}:
    get:
      tags: [Bins]
      summary: Get a single bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bin details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Bins]
      summary: Update a bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBinRequest'
      responses:
        '200':
          description: Bin updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Bins]
      summary: Soft-delete a bin
      description: Sets deleted_at timestamp. Bin can be restored from trash.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bin soft-deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bin deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /bins/{id}/restore:
    post:
      tags: [Bins]
      summary: Restore a soft-deleted bin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bin restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /bins/{id}/permanent:
    delete:
      tags: [Bins]
      summary: Permanently delete a bin
      description: Only works on soft-deleted bins. Removes the bin and its photos from storage.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Bin permanently deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Bin permanently deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /bins/{id}/photos:
    post:
      tags: [Bins]
      summary: Upload a photo to a bin
      description: Accepts JPEG, PNG, or WebP up to 5MB.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        '201':
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /bins/{id}/add-tags:
    put:
      tags: [Bins]
      summary: Add tags to a bin
      description: Merges new tags with existing ones (does not replace).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTagsRequest'
      responses:
        '200':
          description: Tags added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ════════════════════════════════════════════════════════
  # Photos
  # ════════════════════════════════════════════════════════
...
  /shapes/bins:
    get:
      tags: [Shapes]
      summary: Proxy bins shape
      description: Proxies to Electric SQL for real-time bin sync. UUID validation on location_id.
      parameters:
        - name: location_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Shape data stream
        '401':
          $ref: '#/components/responses/Unauthorized'
...